add float overload to correctEndian start moving serialize() and deserialize() code to File class add getter methods in Order, Table classes TrackRow is now a packed structure

+++ libtrackerboy/Module.cpp
@@ -69,7 +69,7 @@ void Module::serialize(std::ofstream &stream) {
 
     auto pos = stream.tellp();
 
-    header.instrumentTableOffset = correctEndian(static_cast<uint8_t>(sizeof(header) + pos));
+    header.instrumentTableOffset = correctEndian(static_cast<uint32_t>(sizeof(header) + pos));
     
     // ignore the next two offsets for now
     auto stoffpos = offsetof(ModuleHeader, songTableOffset) + pos;
